refactor: extract protocol/ — split CoAP wire helpers from DtlsCoapSession
